fix(core): move extraction provenance helper
